convert project to maven
