--- nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=19b3eebd
-build.xml.script.CRC32=0f50d4d7
-build.xml.stylesheet.CRC32=f85dc8f2@1.92.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=19b3eebd
-nbproject/build-impl.xml.script.CRC32=d8a94da0
-nbproject/build-impl.xml.stylesheet.CRC32=3a2fa800@1.92.0.48
