+++ src/main/resources/plugin.yml
@@ -1,0 +1,19 @@
+name: CronPlugin
+version: 0.0.3
+author: AztecMC
+main: org.aztecmc.plugins.cron.CronPlugin
+api-version: 1.13
+commands:
+  cron:
+    description: reload the config and crons
+    permission: cron.reload
+    usage: /<command> reload
+
+permissions:
+  cron.*:
+    description: gives access to all cron commands
+    children:
+      cron.reload: true
+    cron.reload:
+      description: reload the config and crons
+      default: op
